+++ shipyard.py
@@ -36,13 +36,22 @@ import logging
 import logging.handlers
 import os
 try:
+    from os import scandir as scandir
+except ImportError:
+    from scandir import scandir as scandir
+try:
     import pathlib
 except ImportError:
     import pathlib2 as pathlib
 import platform
+try:
+    from shlex import quote as shellquote
+except ImportError:
+    from pipes import quote as shellquote
 import subprocess
 import sys
 import tempfile
+import threading
 import time
 try:
     import urllib.request as urllibreq
@@ -86,6 +95,8 @@ _NVIDIA_DOCKER = {
     }
 }
 _NVIDIA_DRIVER = 'nvidia-driver.run'
+_GLUSTER_VOLUME = '.gluster/gv0'
+_DATA_XFER_METHODS = ('rsync+ssh', 'scp', 'multinode_scp')
 _STORAGEACCOUNT = None
 _STORAGEACCOUNTKEY = None
 _STORAGEACCOUNTEP = None
@@ -710,7 +721,15 @@ def add_pool(batch_client, blob_client, config):
     # create admin user on each node if requested
     add_ssh_user(batch_client, config, nodes)
     # log remote login settings
-    get_remote_login_settings(batch_client, config, nodes)
+    rls = get_remote_login_settings(batch_client, config, nodes)
+    # ingress data if specified
+    try:
+        ingress_files = config[
+            'pool_specification']['transfer_files_on_pool_creation']
+    except KeyError:
+        ingress_files = True
+    if ingress_files:
+        ingress_data(batch_client, config, rls)
 
 
 def _setup_glusterfs(batch_client, blob_client, config, nodes):
@@ -977,6 +996,358 @@ def add_admin_user_to_compute_node(
             raise
 
 
+def _subprocess_with_output(cmd, shell=False, suppress_output=False):
+    # type: (str, bool, bool) -> int
+    """Subprocess command and print output
+    :param str cmd: command line to execute
+    :param bool shell: use shell in Popen
+    :param bool suppress_output: suppress output
+    :rtype: int
+    :return: return code of process
+    """
+    _devnull = open(os.devnull, 'w')
+    if suppress_output:
+        proc = subprocess.Popen(
+            cmd, shell=shell, stdout=_devnull, stderr=subprocess.STDOUT)
+    else:
+        proc = subprocess.Popen(cmd, shell=shell)
+    proc.wait()
+    _devnull.close()
+    return proc.returncode
+
+
+def _subprocess_nowait(cmd, shell=False, suppress_output=False):
+    # type: (str, bool, bool) -> subprocess.Process
+    """Subprocess command and do not wait for subprocess
+    :param str cmd: command line to execute
+    :param bool shell: use shell in Popen
+    :param bool suppress_output: suppress output
+    :rtype: subprocess.Process
+    :return: subprocess process handle
+    """
+    _devnull = open(os.devnull, 'w')
+    if suppress_output:
+        proc = subprocess.Popen(
+            cmd, shell=shell, stdout=_devnull, stderr=subprocess.STDOUT)
+    else:
+        proc = subprocess.Popen(cmd, shell=shell)
+    return proc
+
+
+def _subprocess_wait_all(procs):
+    # type: (list) -> list
+    """Wait for all processes in given list
+    :param list procs: list of processes to wait on
+    :rtype: list
+    :return: list of return codes
+    """
+    rcodes = [None] * len(procs)
+    while True:
+        for i in range(0, len(procs)):
+            if rcodes[i] is None and procs[i].poll() == 0:
+                rcodes[i] = procs[i].returncode
+        if all(x is not None for x in rcodes):
+            break
+        time.sleep(0.03)
+    return rcodes
+
+
+def _scp_data(src, dst, username, ssh_private_key, rls, eo):
+    # type: (str, str, str, pathlib.Path, dict, str, str) -> None
+    """Secure copy data
+    :param str src: source path
+    :param str dst: destination path
+    :param str username: username
+    :param pathlib.Path: ssh private key
+    :param dict rls: remote login settings
+    :param str eo: extra options
+    """
+    recursive = '-r' if pathlib.Path(src).is_dir() else ''
+    _rls = next(iter(rls.values()))
+    ip = _rls.remote_login_ip_address
+    port = _rls.remote_login_port
+    del _rls
+    cmd = ('scp -o StrictHostKeyChecking=no '
+           '-o UserKnownHostsFile=/dev/null -p '
+           '{} {} -i {} -P {} {} {}@{}:"{}"'.format(
+               eo, recursive, ssh_private_key, port, shellquote(src),
+               username, ip, shellquote(dst)))
+    logger.info('begin ingressing data from {} to {}'.format(
+        src, dst))
+    start = datetime.datetime.now()
+    rc = _subprocess_with_output(cmd, shell=True)
+    diff = datetime.datetime.now() - start
+    if rc == 0:
+        logger.info(
+            'finished ingressing data from {0} to {1} in {2:.2f} sec'.format(
+                src, dst, diff.total_seconds()))
+    else:
+        logger.error(
+            'data ingress from {} to {} failed with return code: {}'.format(
+                src, dst, rc))
+
+
+def scantree(path):
+    # type: (str) -> os.DirEntry
+    """Recursively scan a directory tree
+    :param str path: path to scan
+    :rtype: DirEntry
+    :return: DirEntry via generator
+    """
+    for entry in scandir(path):
+        if entry.is_dir(follow_symlinks=True):
+            # due to python2 compat, cannot use yield from here
+            for t in scantree(entry.path):
+                yield t
+        else:
+            yield entry
+
+
+def _multinode_scp_data(src, dst, username, ssh_private_key, rls, eo, mpt):
+    # type: (str, str, str, pathlib.Path, dict, str, str, int) -> None
+    """Secure copy data to multiple destination nodes simultaneously
+    :param str src: source path
+    :param str dst: destination path
+    :param str username: username
+    :param pathlib.Path: ssh private key
+    :param dict rls: remote login settings
+    :param str eo: extra options
+    :param int mpt: max parallel transfers per node
+    """
+    psrc = pathlib.Path(src)
+    if len(rls) == 1 or not psrc.is_dir():
+        _scp_data(src, dst, username, ssh_private_key, rls, eo)
+        return
+    buckets = {}
+    files = {}
+    rcodes = {}
+    for rkey in rls:
+        buckets[rkey] = 0
+        files[rkey] = []
+        rcodes[rkey] = None
+    # walk the directory structure
+    # 1. construct a set of dirs to create on the remote side
+    # 2. binpack files to different nodes
+    total_files = 0
+    dirs = set()
+    for entry in scantree(src):
+        rel = pathlib.Path(entry.path).relative_to(psrc)
+        sparent = str(pathlib.Path(entry.path).relative_to(psrc).parent)
+        if sparent != '.':
+            dirs.add(sparent)
+        if entry.is_file():
+            dstpath = '{}{}/{}'.format(dst, psrc.name, rel)
+            # get key of min bucket values
+            key = min(buckets, key=buckets.get)
+            buckets[key] += entry.stat().st_size
+            files[key].append((entry.path, dstpath))
+            total_files += 1
+    total_size = sum(buckets.values())
+    # create remote directories via ssh
+    logger.debug('creating remote directories: {}'.format(dirs))
+    dirs = ['mkdir -p {}/{}'.format(psrc.name, x) for x in list(dirs)]
+    dirs.insert(0, 'cd {}'.format(dst))
+    _rls = next(iter(rls.values()))
+    ip = _rls.remote_login_ip_address
+    port = _rls.remote_login_port
+    del _rls
+    mkdircmd = ('ssh -o StrictHostKeyChecking=no '
+                '-o UserKnownHostsFile=/dev/null -x '
+                '-i {} -p {} {}@{} {}'.format(
+                    ssh_private_key, port, username, ip,
+                    _wrap_commands_in_shell(dirs)))
+    rc = _subprocess_with_output(mkdircmd.split())
+    if rc == 0:
+        logger.info('remote directories created on {}'.format(dst))
+    else:
+        logger.error('remote directory creation failed')
+        return
+    del ip
+    del port
+    # scp data to multiple nodes simultaneously
+    if mpt is None:
+        mpt = 1
+    logger.info(
+        'ingress data: {0:.4f} MiB in {1} files to transfer, using {2} max '
+        'parallel transfers per node'.format(
+            total_size / 1048576, total_files, mpt))
+    logger.info('begin ingressing data from {} to {}'.format(
+        src, dst))
+    nodekeys = list(buckets.keys())
+    threads = []
+    start = datetime.datetime.now()
+    for i in range(0, len(buckets)):
+        nkey = nodekeys[i]
+        thr = threading.Thread(
+            target=_scp_thread_worker,
+            args=(mpt, nkey, rcodes, files[nkey],
+                  rls[nkey].remote_login_ip_address,
+                  rls[nkey].remote_login_port, username, ssh_private_key, eo)
+        )
+        threads.append(thr)
+        thr.start()
+    for i in range(0, len(buckets)):
+        threads[i].join()
+    diff = datetime.datetime.now() - start
+    success = True
+    for nkey in rcodes:
+        if rcodes[nkey] != 0:
+            logger.error('data ingress failed to node: {}'.format(nkey))
+            success = False
+    if success:
+        logger.info(
+            'finished ingressing {0:.4f} MB of data in {1} files from {2} to '
+            '{3} in {4:.2f} sec ({5:.3f} Mbit/s)'.format(
+                total_size / 1048576, total_files, src, dst,
+                diff.total_seconds(),
+                (total_size * 8 / 1e6) / diff.total_seconds()))
+
+
+def _scp_thread_worker(
+        mpt, node_id, rcodes, files, ip, port, username, ssh_private_key, eo):
+    # type: (int, str, dict, list, str, int, str, pathlib.Path, str) -> None
+    """Worker thread code for secure copy to a node with a file list
+    :param int mpt: max parallel transfers per node
+    :param str node_id: node id
+    :param dict rcodes: return codes dict
+    :param list files: list of files to copy
+    :param str ip: ip address
+    :param int port: port
+    :param str username: username
+    :param pathlib.Path: ssh private key
+    :param str eo: extra options
+    """
+    i = 0
+    while True:
+        procs = []
+        for j in range(i, i + mpt):
+            if j >= len(files):
+                break
+            file = files[j]
+            src = file[0]
+            dst = file[1]
+            cmd = ('scp -o StrictHostKeyChecking=no '
+                   '-o UserKnownHostsFile=/dev/null -p '
+                   '{} -i {} -P {} {} {}@{}:"{}"'.format(
+                       eo, ssh_private_key, port, shellquote(src), username,
+                       ip, shellquote(dst)))
+            procs.append(_subprocess_nowait(cmd, shell=True))
+        rc = _subprocess_wait_all(procs)
+        for _rc in rc:
+            if _rc != 0:
+                logger.error(
+                    'data ingress to {} failed with return code: {}'.format(
+                        node_id, _rc))
+                rcodes[node_id] = _rc
+                return
+        i += len(procs)
+        if i == len(files):
+            break
+    rcodes[node_id] = 0
+
+
+def ingress_data(batch_client, config, rls=None):
+    # type: (batch.BatchServiceClient, dict, dict) -> None
+    """Ingresses data into Azure Batch
+    :param batch_client: The batch client to use.
+    :type batch_client: `batchserviceclient.BatchServiceClient`
+    :param dict config: configuration dict
+    :param dict rls: remote login settings
+    """
+    try:
+        files = config['global_resources']['files']
+    except KeyError:
+        logger.debug('no files to ingress detected')
+        return
+    try:
+        username = config['pool_specification']['ssh']['username']
+    except KeyError:
+        username = None
+    if rls is None:
+        rls = get_remote_login_settings(batch_client, config)
+    for fdict in files:
+        src = fdict['source']
+        try:
+            shared = fdict['destination']['shared_data_volume']
+        except KeyError:
+            shared = None
+        try:
+            storage = fdict['destination']['storage_account_settings']
+        except KeyError:
+            storage = None
+        if shared is not None and storage is not None:
+            raise RuntimeError(
+                'cannot specify both shared data volume and storage for the '
+                'destination for source: {}'.format(src))
+        if shared is not None:
+            if username is None:
+                raise RuntimeError(
+                    'cannot ingress data to shared data volume without a '
+                    'valid ssh user')
+            try:
+                method = fdict['destination']['data_transfer'][
+                    'method'].lower()
+            except KeyError:
+                raise RuntimeError(
+                    'no transfer method specified for data transfer of '
+                    'source: {} to {}'.format(src, shared))
+            try:
+                eo = fdict['destination']['data_transfer']['extra_options']
+                if eo is None:
+                    eo = ''
+            except KeyError:
+                eo = ''
+            try:
+                mpt = fdict['destination']['data_transfer'][
+                    'max_parallel_transfers_per_node']
+                if mpt is not None and mpt <= 0:
+                    mpt = None
+            except KeyError:
+                mpt = None
+            try:
+                ssh_private_key = pathlib.Path(
+                    fdict['destination']['data_transfer']['ssh_private_key'])
+            except KeyError:
+                raise RuntimeError('ssh private key must be specified')
+            if not ssh_private_key.exists():
+                raise RuntimeError(
+                    'ssh private key does not exist at: {}'.format(
+                        ssh_private_key))
+            # convert shared to actual path
+            shared_data_volumes = config['global_resources'][
+                'docker_volumes']['shared_data_volumes']
+            for key in shared_data_volumes:
+                if key == shared:
+                    driver = shared_data_volumes[key]['volume_driver']
+                    break
+            if driver == 'glusterfs':
+                if (config['pool_specification']['offer'].lower() ==
+                        'ubuntuserver'):
+                    dst = '/mnt/batch/tasks/shared/{}/'.format(_GLUSTER_VOLUME)
+                else:
+                    dst = '/mnt/resource/batch/tasks/shared/{}/'.format(
+                        _GLUSTER_VOLUME)
+            else:
+                raise RuntimeError(
+                    'data ingress to {} not supported'.format(driver))
+            if method == 'scp':
+                _scp_data(src, dst, username, ssh_private_key, rls, eo)
+            elif method == 'multinode_scp':
+                _multinode_scp_data(
+                    src, dst, username, ssh_private_key, rls, eo, mpt)
+            elif method == 'rsync+ssh':
+                raise NotImplementedError()
+            else:
+                raise RuntimeError(
+                    'unknown transfer method: {}'.format(method))
+        elif storage is not None:
+            # container = fdict['destination']['container']
+            raise NotImplementedError()
+        else:
+            raise RuntimeError(
+                'invalid file transfer configuration: {}'.format(fdict))
+
+
 def _add_global_resource(
         queue_client, table_client, config, pk, p2pcsd, grtype):
     # type: (azurequeue.QueueService, azuretable.TableService, dict, str,
@@ -1175,6 +1546,38 @@ def _adjust_settings_for_pool_creation(config):
                 'disabling ssh user creation due to script being run '
                 'from Windows and no public key is specified')
             config['pool_specification'].pop('ssh', None)
+    # ensure file transfer settings
+    try:
+        xfer_files_with_pool = config['pool_specification'][
+            'transfer_files_on_pool_creation']
+    except KeyError:
+        xfer_files_with_pool = True
+        config['pool_specification'][
+            'transfer_files_on_pool_creation'] = xfer_files_with_pool
+    try:
+        files = config['global_resources']['files']
+        shared = False
+        for fdict in files:
+            if 'shared_data_volume' in fdict['destination']:
+                shared = True
+                break
+        if _ON_WINDOWS and shared and xfer_files_with_pool:
+            raise RuntimeError(
+                'cannot transfer files to shared data volume on Windows')
+    except KeyError:
+        pass
+    # force disable block for global resources if ingressing data
+    try:
+        block_for_gr = config[
+            'pool_specification']['block_until_all_global_resources_loaded']
+    except KeyError:
+        block_for_gr = True
+    if xfer_files_with_pool and block_for_gr:
+        logger.warning(
+            'disabling block until all global resources loaded with '
+            'transfer files on pool creation enabled')
+        config['pool_specification'][
+            'block_until_all_global_resources_loaded'] = False
 
 
 def resize_pool(batch_client, config):
@@ -1465,8 +1868,9 @@ def add_jobs(batch_client, blob_client, config):
                             'global_resources']['docker_volumes'][
                                 'shared_data_volumes'][key]
                         if dvspec['volume_driver'] == 'glusterfs':
-                            run_opts.append('-v {}:{}'.format(
-                                '$AZ_BATCH_NODE_SHARED_DIR/.gluster/gv0',
+                            run_opts.append('-v {}/{}:{}'.format(
+                                '$AZ_BATCH_NODE_SHARED_DIR',
+                                _GLUSTER_VOLUME,
                                 dvspec['container_path']))
                         else:
                             run_opts.append('-v {}:{}'.format(
@@ -1884,27 +2288,31 @@ def del_all_jobs(batch_client, config):
 
 
 def get_remote_login_settings(batch_client, config, nodes=None):
-    # type: (batch.BatchServiceClient, dict, List[str], bool) -> None
+    # type: (batch.BatchServiceClient, dict, List[str], bool) -> dict
     """Get remote login settings
     :param batch_client: The batch client to use.
     :type batch_client: `batchserviceclient.BatchServiceClient`
     :param dict config: configuration dict
     :param list nodes: list of nodes
+    :rtype: dict
+    :return: dict of node id -> remote login settings
     """
     pool_id = config['pool_specification']['id']
     if nodes is None:
         nodes = batch_client.compute_node.list(pool_id)
+    ret = {}
     for node in nodes:
         rls = batch_client.compute_node.get_remote_login_settings(
             pool_id, node.id)
         logger.info('node {}: ip {} port {}'.format(
             node.id, rls.remote_login_ip_address, rls.remote_login_port))
+        ret[node.id] = rls
+    return ret
 
 
 def stream_file_and_wait_for_task(batch_client, filespec=None):
     # type: (batch.BatchServiceClient, str) -> None
-    """Stream a file and wait for task to complete (streams for a maximum
-    of 30 minutes)
+    """Stream a file and wait for task to complete
     :param batch_client: The batch client to use.
     :type batch_client: `batchserviceclient.BatchServiceClient`
     :param str filespec: filespec (jobid:taskid:filename)
@@ -1947,9 +2355,7 @@ def stream_file_and_wait_for_task(batch_client, filespec=None):
     curr = 0
     end = 0
     completed = False
-    timeout = datetime.timedelta(minutes=30)
-    time_to_timeout_at = datetime.datetime.now() + timeout
-    while datetime.datetime.now() < time_to_timeout_at:
+    while True:
         # get task file properties
         try:
             tfp = batch_client.file.get_node_file_properties_from_task(
@@ -1964,7 +2370,6 @@ def stream_file_and_wait_for_task(batch_client, filespec=None):
         size = int(tfp.response.headers['Content-Length'])
         if size != end and curr != size:
             end = size
-            # get stdout.txt
             frag = batch_client.file.get_from_task(
                 job_id, task_id, file,
                 batchmodels.FileGetFromTaskOptions(
@@ -2319,6 +2724,8 @@ def main():
         get_file_via_task(batch_client, config, args.filespec)
     elif args.action == 'getnodefile':
         get_file_via_node(batch_client, config, args.nodeid)
+    elif args.action == 'ingressdata':
+        ingress_data(batch_client, config)
     elif args.action == 'delstorage':
         delete_storage_containers(
             blob_client, queue_client, table_client, config)